requirements.txt
pandas==1.5.2
python-dotenv==0.21.0
cx_Oracle==8.3.0